+++ packages/moonstone/Scroller/Scroller.js
@@ -178,22 +178,19 @@ class ScrollerBase extends Component {
 			mergedStyle = Object.assign({}, style, {
 				overflowX: props.horizontal,
 				overflowY: props.vertical
-			}),
-			hideNativeScrollbar = !props.hideScrollbars ? css.hideNativeScrollbar : null;
+			});
 
 		delete props.cbScrollTo;
 		delete props.className;
-		delete props.hideScrollbars;
 		delete props.horizontal;
 		delete props.onScrolling;
 		delete props.onScrollStart;
 		delete props.onScrollStop;
-		delete props.positioningOption;
 		delete props.style;
 		delete props.vertical;
 
 		return (
-			<div {...props} ref={this.initRef} className={classNames(className, hideNativeScrollbar)} style={mergedStyle} />
+			<div {...props} ref={this.initRef} className={classNames(className, css.hideNativeScrollbar)} style={mergedStyle} />
 		);
 	}
 }